Separate GUI config and language config
Fixes #234.
