+++ src/main/java/cuchaz/enigma/gui/panels/PanelEditor.java
@@ -2,7 +2,7 @@ package cuchaz.enigma.gui.panels;
 
 import cuchaz.enigma.EnigmaProject;
 import cuchaz.enigma.analysis.EntryReference;
-import cuchaz.enigma.config.Config;
+import cuchaz.enigma.gui.config.GuiConfig;
 import cuchaz.enigma.gui.BrowserCaret;
 import cuchaz.enigma.gui.Gui;
 import cuchaz.enigma.translation.representation.entry.ClassEntry;
@@ -166,6 +166,6 @@ public class PanelEditor extends JEditorPane {
 
 	@Override
 	public Color getCaretColor() {
-		return new Color(Config.getInstance().caretColor);
+		return new Color(GuiConfig.getInstance().caretColor);
 	}
 }
